Fix gUnzip for seemingly multi-gzipped files
Some combinations of user browser, subversion, and nginx result in
gzipped files being delivered to the browser as "double" gzipped.

The old code tried to guess the combinations which required double
unzipping; now we check for the gzip "magic numbers" in the file
header and unzip recursively until we get the original data.