Datafetcher now provides data in columns instead of rows
This was a big complicated change for little benefit;
might revert.

But it does allow individual CSV columns of data to be slung
independently, and that may be a good thing for performance or memory,
although I haven't actually tested that.

This was a precursor to allowing data with non-numeric columns,
especially strings & factors.