feat: Enable Chrome local file browsing
EXPERIMENTAL!

Chrome (and only Chrome!) supports an experimental
"File System Access API" that enables direct reading of
user files, after the user explicitly grants access to
a folder and its contents.

https://wicg.github.io/file-system-access/

This is fully supported in Chrome (no flags necessary)
-- but it will likely never make it into any other browsers.

So, let's support it if the user is on Chrome.