patch-package vuera: deep:true caused huge performance hit on Deck.gl
Vuera is the package which allows React components (e.g. Deck.gl) to
embed seamlessly with a Vue website like SimWrapper.

Unfortunately changes to any attributes passed into a react component
are deep-copied, which results in enormous memory usage and slowdowns
on the CPU-bound main thread. This completely defeats the purpose of
Deck.gl's extensive GPU-based optimizations.

Changing one line to deep:false fixes this problem; Deck.gl is smart
enough to notice when an object's properties change. At least, as far
as I can tell. And we don't use React components for much else, so
even if it causes other problems this change is worth it.