fix(transit): allow viz-pt-*.yaml file to specify correct network (old kelheim)
The transit viewer blindly assumes that the correct network file to load
is named s/transitSchedule/network/, meaning that if the schedule is

kelheim-v1.0-transitSchedule.xml.gz

--> then the network must be kelheim-v1.0-network.xml.gz

This doesn't work if the user has named their network file something else.
In this situation, the user must create a viz-pt-*.yaml file which
defines the filenames to load for both transitSchedule and network.

Example is at
https://simwrapper.github.io/staging/public/de/kelheim/kelheim-v1.0/input/viz-pt-transit.yaml

fixes #172