fix: Firefox (sometimes) fails to load large MATSim XML networks
The fast-xml-parser library sometimes triggers a hard hang on Firefox,
without any error or trap. Just hangs while parsing the XML.

This "fix" introduces a console.log() that adds just enough delay
in the processing to let Firefox not crash. This temporary hack needs
to be here as long as Firefox and fast-xml-parser don't play nice.

fixes: #260