feat: Dashboard subtabs
Dashboards can now have layout SECTIONS on the left and SUBTABS on the top.

Each "dashboard-*.yaml" now defines a dashboard section. The section titles
are along the left side of the dashboard; so the old tabs along the top
have moved to the left-side nav instead.

And in each dashboard-*.yaml file, a new "subtabs" section can include
an array which defines tabs along the top of the section.

Each subtab has its own "layout" section which defines the rows and
card panels in that subtab.

header:
  title: My dashboard

subtabs:
  - subtab: "Mode Split"
    layout:
       row1:
          # .....
       row2:
          # etc...
  - subtab: "Traffic"
    layout:
       row1:
          # ....

So, if you are converting a dashboard to add new subtabs, you
need to create the "subtabs" section, add a subtab, and move the old "layout" to be
a child of the subtab.