fix: Area map using avro-format network doesn't link to CSV data files
Unhelpful rename of key attribute in Avro network from "id" to "linkId"
meant that CSV datasets couldn't properly join.

Now there is a duplicate "id" along with "linkId" so as to not break
any configs using the linkId property; doesn't take extra memory because
it's just a duplicate reference, not a duplicate column.

Also warn the user if they try to join without providing the join column.

fixes: #425