fix(map): shapefile/network filters were disabled; fixed now (dataset filters TBA)
YAML and interactive selection of geojson/shape filters is now working
NOTE - this allows user to filter the SHAPES -- the network/geojson/etc
based on data fields in that shapefile. For example, hide centroid links
in a network file.

Filtering datasets themselves will be added in a future update.

fixes #423