fix: Some improvements to mobile dashboard layouts
SimWrapper is intended as desktop software, but phones do exist so
we are adding some bare-minimum support for loading dashboards on
mobile devices.

Browsers on mobile have very restricted memory and processing
capabilities; and of course the narrow screen sizes present huge
problems for mouse pointer-based UIs. Despite this we are trying
to shoehorn in some minimum usability.

You will still (and will always) have better results on desktop.